Change the LCLS1 DAQ to talk to the LCLS2 logbook.
Fundamentally, change the RunAllocator API and make the rest of the code cater to this new API.
In addition, remove all references to experiment number and use the name instead.
The LCLS1 DAQ seems to pick up it's configuration from a file; this file needs to be changed.

Also, changed the run parameter registration to batch mode for performance reasons.
In the pydaq, the experiment call now returns the experiment name.